+++ easycv/models/detection/utils/__init__.py
@@ -3,9 +3,10 @@
 from .boxes import (batched_nms, bbox2result, bbox_overlaps, bboxes_iou,
                     box_cxcywh_to_xyxy, box_xyxy_to_cxcywh, distance2bbox,
                     fp16_clamp, generalized_box_iou, postprocess)
+from .detr_utils import DetrPostProcess
 from .generator import MlvlPointGenerator
 from .matcher import HungarianMatcher
-from .misc import (DetrPostProcess, accuracy, filter_scores_and_topk,
+from .misc import (accuracy, filter_scores_and_topk,
                    gen_encoder_output_proposals, gen_sineembed_for_position,
                    interpolate, inverse_sigmoid, output_postprocess,
                    select_single_mlvl)
